requirements.txt
yfinance
numpy
pandas
seaborn
matplotlib.pyplot
sklearn
requests
os
time
streamlit